传统方法实现生产者和消费者模式

+++ JavaInterview/src/main/java/com/kxw/thread/trandtion/ProductAndCustomer.java
@@ -62,7 +62,7 @@ class ProductThread implements Runnable{
     public void run() {
         for(int i=0; i<20; i++){
             try {
-                Thread.sleep((long)Math.random()*1000);
+                Thread.sleep((int) (Math.random() * 1000));
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
@@ -83,7 +83,7 @@ class CustomerThread implements Runnable{
     public void run() {
         for(int i=0; i<20; i++){
             try {
-                Thread.sleep((long)Math.random()*1000);
+                Thread.sleep((int) (Math.random() * 1000));
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
